documentService/requirements.txt
django
djangorestframework
packaging
drf-yasg